+++ linux/keymap_se.c
@@ -1,5 +1,5 @@
 // keymap_se.c - keymap - PC Swedish
-const int keymap_se[8][128] = {
+const int16_t keymap_se[8][128] = {
   // base
   { 0,27,'1','2','3','4','5','6','7','8','9','0','+',-3,8,9,
     'q','w','e','r','t','y','u','i','o','p',229,-18,13,0,'a','s',
